Use a shared context for testing overview
As the different sections of the overview are going to be tested
independently from each other, they use a shared context to setup the
tests.

This prevents the setup code from being duplicated

NOTE: the overview.yaml file is not created in the setup. Not all the
tests use the overview.yaml so it needs to be explicitly created.