Generalise the method for creating the table
A similar method shall be used for creating the Domain and Group
overview table. The catch is the namespaces it uses to render and the
fields are different. This is why they need to be passed into the
method.

Otherwise it operates in a similar way as before. As this is a general
method, it does not automatically add the name column. Instead this will
need to be passed in as one of the fields.