Break the Overview::Table into its own file
To keep the command code clear, the Overview::Table has been broken out
into it's own file.

It has been put into the Overview module as the `src` directory is
getting very long.

The specs still share a context as this makes testing the command and
the table easier.

Addresses:
https://github.com/alces-software/metalware/pull/319#discussion_r170565040