Render the bmc username and password for each node
As genders do not have configs (unless they are also a group), the
method of rendering the bmc credentials for them does not work.

Instead the credentials need to be rendered individually for each node.

This allows the `ipmi` command to run on both groups and genders. It
also allows nodes to have different passwords withing the same group.