Add back the create_ipmitool_command method
Turns out other commands extensively make use of this method so it can
not be removed. However as the internals do require the node object, the
other commands needed to be updated so they pass it in.

As the other commands use a different `host` this to needs to be passed
in.