Allow the mocked `alces` to be reset
This is important for the few specs that alter the state of an `alces`
object. This is because of the caching that happens at two locations.

Locations:
1. Internally within an instance of Alces. There is extensive caching
which can not be reliable turned off. This provides the speed benefits
by preventing additional file loads. If a file does change however, the
object becomes stale and needs to be thrown out.

2. Caching within a spec. The use of AlcesUtils creates a test version
of alces which is returned by `Alces.new` instead of a new version. This
allows a specially crafted version to be made in advanced and then used
as a replacement. This works fine in the commands, as there is only ever
a single version of `Alces` (gui excluded).

The issue is when testing a command that changes a file (e.g.
configure). The outcome of the command can not be accessed as the spec
still caches the old version of `alces`. For this reason, a `reset_alces`
command has been added to force a new version to be created.

Due to the dynamic nature, `let!` can not be used. Instead a `alces`
method has been added to the spec. Then `Alces.new` is wrapped to either
return the cached version OR to cache a new version. This allows the
cache to be cleared by `reset_alces`.