Move the create_question method into QuestionTree
This is to reduce the coupling between Configurator and
Configurator::Question. From now on the two will not communicate
directly with each other, they will have to go through QuestionTree.

Eventually Configurator::Question should be removed entirely OR reduced
down to the Highline methods only. This will allow the removal of the
`make_question` method. It can then be replace with either `question`
or `ask_question` method directly on the `QuestionTree` object.

The identifier is also always converted to a symbol. It is always used
as a symbol so `QuestionTree` should be responsible for its conversion.
During the `Validation` it is a string as that is how it appears in
`configure.yaml`.

Also do not assume that the `content` will always be a OpenStruct. This
should be the case, however some of the specs violate this.