Delegate the parameters instead of assigning them
As the `Question` object is essentially wrapping the `QuestionTree`
object, most of the parameters can be delegated to. The two classes are
still coupled, however less now that the `properties` object has been
removed.

Properties was an alias to content, and remember DTTSC

For this to work, the `content` hash within `QuestionTree` has been
wrapped to be an OpenStruct. This is the logical place to do it as it is
where it gets used. The `Validation` can be adjust so it doesn't wrap
the content anymore.