Always save the raw answer into the question tree
The answer needs to be saved because it used by the conditional
questions. Before a conditional question is asked, it checks if its
parent has a truthy answer.

For this reason, the `Question` object can save the answer directly.
This removes the need for `Configurator` to do it on its behalf.