Make ipmi commands handle individual failures (fixes #337)
Previously if any individual `ipmitool` command run within a `metal
ipmi` or `metal power` command failed, then the whole command would blow
up and give an error.

This commit fixes this so we show stdout for `ipmitool` commands which
succeed, and stderr for `ipmitool` commands which fail, and don't blow
up entirely in either case.