Simplify the iterator logic to the most reduced form
Previously a big `if` statement was used to determine if an answer needs
to be saved. However most of its logic was obsolete to the point that it
was never used.

The comment has also been removed because it is obviously not applicable
to what is going on.

This is the most reduced state that all the tests work