Add method onto QuestionTree to find the default answers
Previously each level of questions could have a different default.
However this was untested and didn't work and thus was removed.

From now on, each identifier maps to a set default regardless of level.
Currently this mapping only considers defaults set at the domain level.

In the future, the defaults for the other levels can be merged in.
NOTE: Whilst the other level defaults maybe merged in future, the same
merge will occur regardless of which level is currently in scope.