Update spec to test saving of inherited answers
Previously there was no spec at all for how inherited answers should be
saved. This possible has lead to the current bugs in the configure
process (It is hard to tell for certain as there is no spec to compare
against).

This commit updates the spec to include test on inherited answers
against a wide range of edge cases concerning the default.

It also removes the save condition from within the configure loop. This
was done for two reasons:
1. It created more errors than it solved,
2. It should be done within the save method.

This does break a couple of the specs which are currently being skipped.
They will be fixed in the next commits.

How the spec asks the questions also had to be updated. Previously the
input file was completely `rewind`ed. This meant if the method was used
twice, the first answers would be read twice. Instead it is only rewound
the number of characters that we actually wrote. This allows the same
input stream to be used multiple times.

Also updated LoadSaveBase to use `name` rather than `file` as the local
variable to make it more consistent.