Remove the unused 'old_answer' method
The only place the old_answer was being used was within the spec.
However the spec shouldn't be testing the `old_answer` like that as it
is duplicating the purpose of the namespace and answer has merging.

Instead the spec should focus on the final saved answer instead. Which
it then passes.