Don't save domain answers if they match yaml defaults
This commit fixes the answer saving for domain level questions. They do
not get saved if they match the root level defaults from
`configure.yaml`.

This does not fix the bug for the other sections.

The node_spec needed to be updated for this change to work. Previously
no `nodeattr` commands where being called so everything worked fine.
However now `nodeattr` is required and thus needs to be spoofed.