Do not save group answers if it matches the domain's
This is because group answers inherit off the domain and thus do not
need to be saved.

For the spec to work correctly, `alces` needs to be reset after each
configure. This is because configuring changes the answer files which
have already been cached.

Because `alces` is reset after each configure, it does not need to be
reset within the specs themselves.