Ensure the un-rendered answers are used
The configuration process should only deal with unrendered answer as
that is what gets saved. The rendered answers can not always be rendered
within the context that they are set.

Besides this point, they also will affect the equality operator that
determines if the answer is saved.

An easy fix is to convert the answer object into a hash, which contains
the unrendered answers.