Add back the old command inputs to `ipmi`
The `ipmi` command as gone through multiple iterations in which the
command input was originally meant to wrap the underling bash command.

Originally it meant that `-c` was used to specify the input. When this
was ported into ruby, the command input become `-k` and `--command`. The
reason being `-c` was taken by the now defunct `--config` input. There
was also issues with implicit short tags that are now resolved.

Following the refactor to allow `ipmi` to work on cloud nodes, the input
was unified and the tags where removed all together.

This has caused a compatibility nightmare across the last 5 versions of
metalware as the `cli` keeps changing.

To resolve this, the `-c`, `-k`, and `--command` tags have been added
back into the command. The tags are used to support the legacy syntax
only. They are otherwise ignored by the `ipmi` command itself.