Create Command class to parse IPMI input
Now that the ipmi command input can come from two places, it is easier
to create a class to handle the parsing. This allows the parsing to be
tested independently from IPMI itself.

If both inputs are included, then an error is raised. Otherwise which
ever input is used becomes the command. Similarly it errors if no
command is given.