Run the editor using a standard system call
The editor can not be opened with `capture3` like the other commands.
This would capture stdout and thus prevent the editor from being
displayed. Instead `system` is used as runs the command directly.

The spec has been slimmed down as the editor opening during it causes a
few issues. It causes the terminal to have a fit and has issues with
creating swap files.

As these issues are due to the spec not the codebase itself, the testing
has been reduced somewhat.

Instead of actually running the command, the spec expects
`SystemCommand` to receive the correct method and arguments instead.