Do not turn FakeFS when mocking `nodeattr` commands
By turning `FakeFS` off, it makes `AlcesUtils` non-thread safe. This
only ever causes an issue in the tests, but it is highly likely the
cause of the recent intermittent faults in the build process.

As the build process it Threaded, it is possible for one thread to try
and delete a file whilst another is polling the genders file. If this
happens, FakeFS will be off and it will be as if the file doesn't exist.

The fix is to use bash commands when mocking `nodeattr` as they will
always use the real filesystem. This means `FakeFS` can be left on for
all the other threads.