Create AssetArray object for loading asset files
The assets must be accessible from the namespace. However due to the
large number of predicted asset files, they should not be loaded up
front.

Instead the asset files should only be loaded when they are actually
used. For this reason, a new enumerable `AssetArray` class has been
created. It uses a dedicated `[]` method that loads the assets.

Internally it stores an array of `files`. Note this array is generated
on `initialize` and thus can not be modified after the fact.

The `each` method has not been defined yet, this will be added at a
future date.