Allow any heredoc delimiter and a few extract arguments
The use of `EOF` as a delimiter is extensive and isn't going to change
anytime soon. Instead, the cop has been turned off.

Also `a`, `s`, `b`, `e`, `_a`, `_b`, and `_e` are now allowed as
argument inputs. Technically they are two short and thus where being
flagged. However they are used to denote the inputs to `method_missing`
and as an error input.