Update asset commands to include argument inputs
The argument for the `edit` command is the name of the asset itself.
This enough to locate the asset file and trigger the edit process.

In the case of `add` command, an `ASSET_TYPE` is also required. The
type is used to locate the template, which is then saved according to
the `ASSET_NAME`.