Break out the tempfile logic from `open_copy`
The `open_copy` method is quite complex due to the amount of file
handling that needs to be preformed.

This can be broken out into two categories:
1. Handling the source, validation, and destination files,
2. Creating and deleting the temporary file.

To simplify the method, the code associated with temporary file has been
broken out into another method. Only the file path is required by the
actual utility