Add a optional node input that assigns the node
This optional node input can be included with both the `add` and `edit`
asset commands. This will trigger the asset to be assigned to the node.

A shared_example is used for the spec so it can be easily used by both
commands. Currently only the `edit` command implements it however.