Updated MetalROS so it can be inherited from
During the convert_value method, there was a hard coded reference to
MetalRecursiveOpenStruct. This meant the class could not be inherited
from without changing the data structure when the recursion occurred.

This has been updated to refer to `self.class` so the data structure
class remains constant.

Also the `template_block` variable has been renamed
`convert_string_block` as it may not be templating. Instead it is the
block that is called when ever a string is to be returned.