Remove the block input to Asset namespace
The asset namespace does not require a block. MetalROS have been
designed to take a templating block and to pass it down recursively.

The `Asset` namespace is different as it will use the block to provide
the linking between assets. This means the block is the same regardless
of how and when it is initialized.

This means the block input needs to be removed. However as MetalROS will
continue providing the block recursively, the class must not error if it
is provided.