Ensure the Asset namespace does not render strings
The Asset namespace must not render string if a render block is
provided. There are a few ways for ruby magic to pass the templating
block down and thus it is possible that strings might get rendered.

This should never be the case, even if the `Asset` namespace is
initialized with a templating block.