Remove Asset namespace and merge with AssetArray
It will not be possible to make a dedicated Asset namespace class.
In order for assets to reference each other, the Asset array needs
access to `alces.asset`. However the underlining `MetalROS` does not
store `alces` and thus will not know how to access the assets without
reloading the files.

The only fix is to provide an external block to `MetalROS` which
contains `alces`. This block will then be recursively passed down and
thus allows access to the other assets.

Thus the `Asset` namespace has been removed. Instead the code for
referencing other assets will be included within the `AssetArray` class.