Load the asset data as a MetalROS
The `MetalRecursiveOpenStruct` will assets to reference each other.
However the spec needed to be updated to accommodate its use.

Particularly, there is no equality operator for `MetalROS` as the use of
blocks for rendering make it impossible to compare to objects. Thus they
are only equal if they are the same object.

This means the specs need to convert the `MetalROS` to a hash before
comparing the data.