Load the asset metadata into the hash
The assets metadata is loaded into the hash. This currently contains the
assets name, but might in future contain the type as well.

The inclusion of the `metadata` key has caused some of the existing
tests to break. Some of the array specs have been updated, to include
the metadata hash in the comparison.

However the bulk of the specs have been updated to only check if the
asset includes the initial keys, but ignore the additional metadata key.