Test build files with the correct data structure
The build files render has silently broken due to a miss match
between how it is tested and actually used.

RuboCop was used to replace `each { |_k, value| ... }` with the
`each_value { v ... }` syntax. This works fine for a Hash however
MetalRecursiveOpenStruct does not implement `each_value` and
returns nill instead.

However the build method was being tested with a `Hash` and thus
passed. The test has now been updated to use MetalROS.

The bug was fixed by adding a `each_value` method to MetalROS.

The initialization of MetalROS has also been updated so it
includes a default string conversion block that just returns the
raw string. This should make using it in the tests easier.