Replace FilePath.asset with Record::Path.asset
To easy the transition to storing assets in sub directories, all
existing calls to FilePath.asset have been changed to the new
Record::Path.asset method.

Some of these methods may need to be changed back in the future,
paticularly those concerned with the creation of a new asset. However
the starting point should be the use of the new method.