Implement method for finding existing asset paths
The `Records::Path.asset` method will only find paths for assets that
exist. This is currently a trivial tasks because a flat directory
structure is used to store the assets. However this becomes
significantly harder once the assets are stored in sub-directories by
type.

To facilitate this switch, the current `Records::Path.asset` method
supports both file structures. This will need to be updated in the
future.

Because `Records::Path.asset` only works if the asset exists, it causes
issue with the previous commit that replaces the use of FilePath.

Firstly, this will not work with any thing creating an asset. So
AlcesUtils and the add command needed to be converted back to using
FilePath.

This is particularly apparent in the specs, which where manually
generating paths and then dumping the content. Instead they now use the
updated version of AlcesUtils which handles the paths form them.

Finally the `edit`, `delete` and `link` command needed an updated
concerning the use of the `error_if_asset_is_missing`. With this
update, a `nil` is returned if the asset is missing.

A new `asset_path_or_error_if_missing` method now combines finding the
asset path and error if it is missing.