Handle any methods that need to be inherited by record commands
This file is intended to replace CommandHelpers::AssetHelper. Instead of
being a mixin this is inherited by the record commands which will
utilise the same methods as before but in a slightly different format.

Currently this is only used by the asset commands but if and when commands
are created for other forms of record this class can be extended to
support them.