Combined asset mixins into single concern
Now that the `AssetHelper` has been removed, it has become apparent
that there is still a role for an asset mixin.

So instead of having two separate mixins for the AssetCache and the
EnsureAssetExists, they have been combined into a single concern.

Also the inheritance on `post_setup` has been removed. It is cleaner if
commands that require the asset to exist make the call manually.