Move missing asset error into Records::Path
Instead of have a separate method in a concern that errors if an asset
is missing, the error check can occur in Records::Path.

A `missing_error` flag has been added to the `asset` method which will
trigger a error to be raised if an asset is missing instead of
returning nil.