Store assets in sub-directories by type
This encodes the assets type in the file path which can then be loaded
into the metadata at a future date.

The directory names themselves are pluralized. This way when they are
loaded into the namespace, it will be easier to use the following
syntax:
`assets.pdus.each` OR `assets.pdus.pdu1`

This commit required updating `FilePath.asset` to take two inputs (the
types dirname and asset name). As such all its uses had to be updated as
well.