Rename `Records::Path` to `Records::Asset`
With the introduction of the `available` method, the class is becoming
increasingly tied to assets themselves. Originally `layout` methods
however this should probably be in a separate class all together and use
inheritance to share the common aspects.

As such `Records::Path.asset` has been renamed `Records::Asset.path`.
This makes the `Records::Asset.available?` more readable as well.