Make all AssetArray methods reserved asset names
The methods on `AssetArray` can not be used as asset names, otherwise
they can not be loaded into the namespace. Due to the number of methods,
it is hard to know which terms are reserved in advance.

Hence it is better if the `available?` method does this check instead of
`AssetArray`.