Ensure sub-AssetArray's do not have type methods
The purpose for creating an asset array from a set of predefined
loaders, is so sub asset arrays can be created.

Eventually the main AssetArray created in the namespace will contain
"type" methods (e.g. `racks`) which will return an AssetArray of racks
only (aka a sub asset array).

As the sub asset array all shares a type, they do not need the type
methods to be defined on them. This commit ensures that is the case.