Define the types method on the main AssetArray
If an `AssetArray` is initialized without a set of loaders, it means it
is the main array. As such it needs to contain all the `type` methods so
the assets can be grouped by type.

This commit defines the methods. Note that sub arrays do not define
these methods as tested in a previous commit.