Return a sub AssetArray from the type methods
The sub array will eventually be initialized with all the loaders that
match the asset type. For the time being, the asset arrays are empty.

However this commit also implements caching of the sub array. The sub
array is being created during the initialization process and stored as
an instance variable.

This means the sub array is only ever created once and then cached
internally within the asset. This will cut down on how often the array
has to be searched to find its assets.