Implement a utility for finding either an asset type or layout by name
The asset add command needs to be able to create an asset from either a
type or a layout. The base_path method (name subject to change) takes a given
name and searches for it within the list of types and saved layouts.
Returning the it's respective path if it is found or raising an error if it
doesn't exist.

This is then to be used as the base of the asset that is being created.