Extract similar code to Records::Record
Both the layout and asset add commands need to error if there respective
records already exist. Due to the similarity between the methods, it is
ideal if the code was extracted in to a single place.

Thus the code has been moved into Records::Record as that is where most
of the data it requires comes from anyway.

This means any changes in the warning will be duplicated between the
commands.