Move `path_with_types` into AssetBuilder
This is the more appropriate location for the method as it is where the
path will ultimately get used.

Also the method shouldn't error as it will eventually be used within a
loop that shouldn't exit early.

Finally the method can be simplified as the path can be found using the
`Records::Layout.path` method and hence doesn't need to be
reconstructed.

The spec for `path_with_types` has been removed as the behaviour will
change in `AssetBuilder`.