Allow the output to stderr to be toggled
By suppressing all the output to `stderr` it makes it hard to use the
inbuilt `rspec` matches.

Instead, rspec suppresses the warnings by default, however there is an
option to turn them back on if need be. This makes it a lot easier to
test if `stderr` receives the correct warning.