Include the type in the Struct as a placeholder
Eventually the asset type will be included. However at this point in
time the code to do this isn't part of this branch.

Thus the type code will be added in the future.

Also a context has been added around the layout exists code. This will
make it easier to add the types spec in the future.