Add new methods to handle finding a type from a given path
Added the record_dir method that needs to be defined in the inherited class. This method
returns the directory of records for the respective class e.g. the Asset class will
return '/v/l/m/assets'.

Added the type_from_path method that returns the type of the path given to it. After
first checking the path contains record_dir for the class calling the method.