Implement `save` asset method
The `save` asset method will automatically copy the asset from its
source to its destination without editing it.

However if the source is invalid it will raise an error. This shouldn't
happen often as it would indicate that a layout is invalid. This should
only occur if the layout has been manually added.

The `save` method uses the `copy_via_temp_file` utility as it will need
to transform the keys in the file eventually.