Convert sub asset notation to sub asset name
The key feature of layouts is its ability to generate sub assets. This
is done by including a string in the format: `type_layout^sub_name`.

When the asset is added, the asset is parsed for all strings matching
that format. This will eventually trigger those sub assets to be made.

This commit implements the conversion of those sub assets to there final
name. Thus the string needs to be replaced with:
`^#{parent_asset_name}_#{sub_name}`

A bit of code was required to traverse the hash as it is a combination
of assets and arrays. To prevent additional file loads, the validation
returns the data if it is valid. This works as it will always be a Hash
and thus truthy