Use AssetBuilder in `asset add` commands
This commit switches `asset add` from using the old `open_and_edit`
method to using the new `AssetBuilder`.

The reason the asset needs to be pushed and then poped is so the sub
asseting is activated.

NOTE: The sub asset will not be made by this commit, however they will
be queued