Remove extraneous setup in `asset add`
The `asset add` command no longer operates like the old `RecordEditor`
commands. Instead it uses the `AssetBuilder` method to add new assets.

This means the path setup can be removed from the command. The error
check if the `type` is missing also needs to be removed as it no longer
applies. The type can be missing as a layout can be used instead.

If the `layout` is missing, the `AssetBuilder` will issue a warning.