Create test suite for sub asseting with a command
Eventually the user will be asked if they wish to edit each asset before
saving it.

To test this, a setup using a layout with a sub asset has been created.
If the user answers "no"/ `false`, than the editor should only open
once.

Currently this is the case as the sub asset hasn't been created at all.