Fix missing layout spec for assets and layouts
Previously the shared example didn't work because an error is raised in
`layout add` but a warning is issued in `asset add`.

The fix is to remove the spec from the shared example and alter it to
work in the above to cases.

The reason for the difference is `asset add` uses the `AssetBuilder`
that is designed to work with warnings. However `layout add` has to
raise an error so it can exit the command early.