Switch Kickstart BuildMethod to use a vanilla render
Previously it used the more complicated `render_to_file` that also saved
the content immediately.

This however is an exception to the rule, as most templates go through
the staging process.

As the template isn't validated, the `render_to_file` method is required
as the file save can occur outside of metalware.