Join the dead threads NOT the alive
Turns out `delete_if` and `delete` return different values. `delete`
returns the deleted values, where `delete_if` returns the original
object.

This meant the alive threads where being joined instead of the dead
ones. This was what caused the build to hang for `local`.

Also the `BuildEvent` needed to be updated to use `loop while` instead
of `loop if`. I don't know what `loop if` does, but what ever it is, it
causes the tests to fail.