Do not pass `input` into the BuildFileRetriever
Currently the caching is a mess. BuildFileRetriever determines the file
path to the cached file, which is then passed to the `Input::Cache`
class that caches the network request. All of which is made by the
`BuildFileRetrievers::Cache` class.

Instead `BuildFileRetrievers::Cache` will take the roll of determining
the file path, and caching the file.

The first stage is to pass `BuildFileRetrievers::Cache` into to the
`BuildFileRetrievers` and then delegate the `download` method.