Move caching out of the Input class
The caching has been moved out of the Input class and into
BuildFilesRetriever::Cache.

The `url` is hashed to give the location it should be saved. This
means it is still possible to find the cached template by name, but it
also allows different files to have the same base name.

Also the BRF::Cache class is solely responsible for Caching. Instead
of the name being set in the original BRF class but the cache in Input.