Test correct comment char used for managed file markers
Extract shared examples for testing managed file updating by Staging,
and have this check that the correct comment character is used when
updating this file. This is in preparation for allowing customizing the
comment character used for managed files, which will be added shortly
and use the same shared examples but expect the comment character
specified.