Switch NetworkAdapter asset to use new format
The interface is a property of a particular node's connection to a
network, via a network adapter port, rather than a property of the port
itself.

Therefore we have decided to store network adapter ports as an array
rather than a hash including the interface, and the canonical source for
interface names will be in each node's network configs.