Restore custom to_json methods from #430
These were originally added in
https://github.com/alces-software/metalware/pull/430, and were then
removed when this was reverted.

Although the changes to the View command in this PR were moved to the
Eval command instead (see
https://github.com/alces-software/metalware/pull/435), and the other
changes made in this PR were then mostly unneeded and reverted, this
particular change is still useful as it means both `eval` and `view`
give more useful/detailed results when assets and asset arrays are
inspected.