Use correct build method if config value is string (fixes #439)
Since the migration to the Node namespace, the selection of the correct
build method was broken when the `build_method` config value was
specified as a string, which is likely most/all of the time since this
comes from human editable YAML config files.

See
https://github.com/alces-software/metalware/commit/e01b30ef39ff9df3ef9c8032fba8e8e685757068#r30002376
for the change and full details of how this stopped working.