Define new method to determine node build method
Define new class method which lives on the BuildMethods class to
create the appropriate build method object given a Node namespace. This
has been re-TDDed to ensure full, clear test coverage of this.

The reason for creating this is that the Metalware namespaces are being
moved to a new project (Underware), but dealing with build methods is
specifically a Metalware concern, so we want to invert where these are
created so this can continue living in Metalware when the namespaces
themselves move.