Add method to render template file
As with `render_string`, these consist of a method on the `alces`
namespace to handle the actual behaviour, and a simple wrapper in the
child namespaces which calls this. The behaviour itself comes from
`src/templater.rb`, and is being added here to allow us to handle
rendering template files in a consistent way.