Remove `hunter` data-specific handling
`hunter` data will now be written to and read from the new `data` part
of the Metalware namespace, rather than a special `hunter` section
specifically for it (which has now been removed).

This is needed as part of the Metalware/Underware split: `hunter` is a
Metalware-specific concept, but when templating we will be using the
Underware namespace, therefore we want to remove knowledge of `hunter`
from this so that Underware doesn't need to know about Metalware
concepts.