Hack prototype GUI out of Metalware
The prototype Metalware GUI was a useful experiment but hasn't been used
for a year or so, almost certainly no longer works following the many
changes made since then, and is unlikely to be wanted again in its
current form.

Rather than keeping the GUI and the code needed to hook Metalware itself
in to this around in its current form, crufting up the codebase while
remaining unneeded and unworking, let's just remove it for now.

If we ever need to refer to the Metalware GUI as first implemented we
can always go back to the point in the history where this was done and
see it working in its original form.

Meanwhile if we ever want a more production-ready Metalware GUI, we will
likely want to implement this from the ground up, taking account of the
many changes made to Metalware since the prototype, the lessons learnt
from building this, and eliminating the original hacks needed to get
this working.