Lock down Underware version Metalware requires
We want Underware to be installed at a sibling directory to Metalware
(so it can be used as an entirely independent tool in its own right),
but also want to lock down the Underware version so we fail fast if the
installed Underware in this directory is not what we require. The
required version of Underware is loaded from a file so we can easily
have this version installed by Metalware itself in production (or CI) if
needed.