Move stubbing constant within test file
This stubbing previously used a method from `Underware::AlcesUtils`,
which stubs a Metalware constant and is only ever used in this one file
in Metalware.  Rather than this convoluted way of doing things, it's
much clearer to just add a method to do the stubbing in the single place
it's needed.