Determine hexadecimal IPs in build method classes
Rather than getting this from the node namespace. Doing this within
Metalware, and in particular in the Kickstart build methods, is a better
location for this than in Underware: firstly as this is a
Metalware-specific concern, and is only used when determining Kickstart
file names, and secondly as the node namespace method will soon be
removed from Underware to resolve
https://github.com/alces-software/underware/issues/3.

This is done by moving the code from the Underware namespace to do this
into Metalware, as well as adding back the `PATH` modification we used
to do so that the mock `gethostip` script, once used to allow
determining a hexadecimal IP value to work for all nodes in tests, is
used again (this had become unused, and the `hexadecimal_ip` method from
Underware's `AlcesUtils` was being used for similar mocking instead).